third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-html-2-expected.txt

Verifies JavaScript pretty-printing functionality.


Running: testDocType
====== 8< ------
<!DOCTYPE HTML>
<body>hello, world</body>

------ >8 ======
Correct mapping for <<body>>
Correct mapping for <hello>
Correct mapping for <</body>>

Running: testComment
====== 8< ------
<!-- comment 1 -->
<html>
    <!-- comment 2-->
    <meta/>
    <body>
        <!-- comment 3-->
        <a>link</a>
    </body>
</html>

------ >8 ======
Correct mapping for <<body>>
Correct mapping for <comment 1>
Correct mapping for <comment 2>
Correct mapping for <comment 3>
Correct mapping for <link>

Running: testNonJavascriptScriptTag
====== 8< ------
<div>
    <script type='text/K'>
        2_&{&/x!/:2_!x}'!R
    </script>
</div>

------ >8 ======
Correct mapping for <type>
Correct mapping for <R>
Correct mapping for <</div>>
Correct mapping for <</script>>

Running: testList
====== 8< ------
<ul>
    <li>foo
    <li>
        hello <b>world</b>
        !
    <li>
        hello <b>world</b>
        <b>i'm here</b>
    <li>bar
    <li>baz
    <li>
        hello <b>world</b>
    <li>another
</ul>

------ >8 ======
Correct mapping for <foo>
Correct mapping for <bar>
Correct mapping for <baz>
Correct mapping for <hello>
Correct mapping for <world>
Correct mapping for <another>

Running: testAutomaticClosingTags
====== 8< ------
<a>
    aaaa
    <b>
        bbbb1
        <c>
            cccc<d>dddd
        </c>
        bbbb2
</a>

------ >8 ======
Correct mapping for <aaaa>
Correct mapping for <bbbb1>
Correct mapping for <bbbb2>
Correct mapping for <cccc>
Correct mapping for <dddd>
